feat(new-fonts-and-css): Add robot and baloo 2 fonts and include shadow box on cafe description
This commit includes a two new fontes. The default for this projet is to use baloo 2 for <h1> and <h2> tags and roboto for the rest. This is done to make the fonts more readable and to avoid the use of the default fonts. The fonts are also used for the title and the subtitle.

The fonts are loaded as metatags on main layout.tsx.

Author: Gabriel Borin Macedo (github : borin98)